+++ core/src/main/java/io/github/scroojalix/npcmanager/utils/storage/StorageFactory.java
@@ -6,7 +6,7 @@ import io.github.scroojalix.npcmanager.NPCMain;
 import io.github.scroojalix.npcmanager.utils.storage.implementation.JsonStorage;
 import io.github.scroojalix.npcmanager.utils.storage.implementation.MongoStorage;
 import io.github.scroojalix.npcmanager.utils.storage.implementation.MySQLStorage;
-import io.github.scroojalix.npcmanager.utils.storage.implementation.StorageImplementation;
+import io.github.scroojalix.npcmanager.utils.storage.implementation.interfaces.StorageImplementation;
 
 public class StorageFactory {