+++ core/src/main/java/io/github/scroojalix/npcmanager/utils/storage/implementation/interfaces/StorageImplementation.java
@@ -1,16 +1,10 @@
-package io.github.scroojalix.npcmanager.utils.storage.implementation;
+package io.github.scroojalix.npcmanager.utils.storage.implementation.interfaces;
 
 import io.github.scroojalix.npcmanager.utils.npc.NPCData;
 
 public interface StorageImplementation {
 
     public String getImplementationName();
-
-    public boolean isRemote();
-    
-    public void init() throws Throwable;
-
-    public void shutdown() throws Throwable;
 
     public void saveNPC(NPCData data) throws Throwable;