+++ CHANGELOG.md
@@ -9,7 +9,7 @@ This project uses the versioning format {phase}.{major}.{minor}, where phase is 
 - Dependency Manager that downloads dependencies into a cache folder and remaps them for use within the plugin. Will be used to implement more storage methods.
 - `MongoDB` as a new storage method.
 ### Changed
-- Name of temp folder for MySQL storage renamed from `temp` to `mysql-temp` so more storage methods can have temp storage.
+- Better temp storage system. Each storage type has its own temp storage folder.
 
 ## [1.4.2] - 2021-1-10
 ### Added